Mensagem curta e clara sobre a mudança feita
