+++ Cadastro/cadastro.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Cadastro - Rupper's Burguer</title>
+    <link rel="stylesheet" href="./cadastro.css">
+</head>
+<body class="auth-page">
+<img src="../assets/img/hamburguerinicial.jpg" alt="imagem Hamburger">
+
+   <header class="header">
+        <div class="container">
+            <div class="header-content">
+                <h1 class="logo">Rupper's Burguer</h1>
+                </div>
+            </div>
+        </div>
+    </header>
+
+    <div class="auth-container">
+        <h1 class="auth-logo">Crie sua Conta</h1>
+        <form class="form-group" action="cadastro.html" method="POST">
+                <label for="nome_completo">Nome Completo</label>
+                <input type="text" id="nome_completo" name="nome_completo" placeholder="Seu nome" class="form-control" required>
+                <label for="email">Email</label>
+                <input type="email" id="email" name="email" placeholder="Usuario@gmail.com" class="form-control" required>
+                <label for="cpf">CPF</label>
+                <input type="text" id="cpf" name="cpf" class="form-control" placeholder="000.000.000-00" required>
+                <label for="telefone">Telefone / WhatsApp</label>
+                <input type="text" id="telefone" name="telefone" class="form-control" placeholder="(00) 90000-0000">
+                <label for="cep">CEP</label>
+                <input type="text" id="cep" name="cep" class="form-control" placeholder="00000-000" required>
+                <label for="endereco">Endereço Completo</label>
+                <input type="text" id="endereco" name="endereco" class="form-control" placeholder="Rua, Número" >
+                <label for="senha">Senha</label>
+                <input type="password" id="senha" name="senha" class="form-control" required>
+                <label for="confirmar_senha">Confirmar Senha</label>
+                <input type="password" id="confirmar_senha" name="confirmar_senha" class="form-control" required>
+            <button type="submit" class="btn-submit">Finalizar Cadastro</button>
+        </form>
+
+        <div class="auth-footer">
+            <p>Já tem uma conta? <a href="../Login/login.html">Faça o login</a></p>
+        </div>
+    </div>
+    <script src="../js/formatacao.js"></script>
+</body>
+</html>
